test: mocking solution improvements
