+++ package.json
@@ -39,6 +39,7 @@
     "@graphql-codegen/typescript-resolvers": "^4.0.6",
     "@graphql-typed-document-node/core": "^3.2.0",
     "@parcel/watcher": "^2.4.1",
+    "@rollup/plugin-graphql": "^2.0.4",
     "@swc-node/register": "^1.9.0",
     "@swc/core": "^1.4.17",
     "@swc/jest": "^0.2.36",
