+++ src/graphql.d.ts
@@ -1,0 +1,9 @@
+declare module "*.graphql" {
+  const Query: import("graphql").DocumentNode;
+  export default Query;
+  export const _queries: Record<string, import("graphql").DocumentNode>;
+  export const _fragments: Record<
+    string,
+    import("graphql").FragmentDefinitionNode
+  >;
+}
