+++ src/test/mocks/fixtures/Pokemons/byOffset/index.ts
@@ -1,0 +1,7 @@
+export * from "./offset0";
+export * from "./offset9";
+export * from "./offset18";
+export * from "./offset27";
+export * from "./offset36";
+export * from "./offset45";
+export * from "./offset54";
