+++ src/test/mocks/fixtures/Pokemons/byOffset/offset27.ts
@@ -1,0 +1,182 @@
+export const offset27 = {
+  pokemons: {
+    count: 1302,
+    next: "http://localhost/api/v2/pokemon/?limit=9&offset=36",
+    previous: "http://localhost/api/v2/pokemon/?limit=9&offset=18",
+    nextOffset: 36,
+    prevOffset: 18,
+    params: {
+      limit: 9,
+      offset: 27,
+    },
+    results: [
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/28.png",
+        id: 28,
+        name: "sandslash",
+        url: "http://localhost/api/v2/pokemon/28/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/28.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/28.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/29.png",
+        id: 29,
+        name: "nidoran-f",
+        url: "http://localhost/api/v2/pokemon/29/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/29.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/29.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/30.png",
+        id: 30,
+        name: "nidorina",
+        url: "http://localhost/api/v2/pokemon/30/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/30.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/30.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/31.png",
+        id: 31,
+        name: "nidoqueen",
+        url: "http://localhost/api/v2/pokemon/31/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/31.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/31.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/32.png",
+        id: 32,
+        name: "nidoran-m",
+        url: "http://localhost/api/v2/pokemon/32/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/32.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/32.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/33.png",
+        id: 33,
+        name: "nidorino",
+        url: "http://localhost/api/v2/pokemon/33/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/33.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/33.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/34.png",
+        id: 34,
+        name: "nidoking",
+        url: "http://localhost/api/v2/pokemon/34/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/34.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/34.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/35.png",
+        id: 35,
+        name: "clefairy",
+        url: "http://localhost/api/v2/pokemon/35/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/35.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/35.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/36.png",
+        id: 36,
+        name: "clefable",
+        url: "http://localhost/api/v2/pokemon/36/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/36.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/36.svg",
+        __typename: "PokemonItem",
+      },
+    ],
+    extended_results: [
+      {
+        id: 28,
+        weight: 295,
+        height: 10,
+        __typename: "Pokemon",
+      },
+      {
+        id: 29,
+        weight: 70,
+        height: 4,
+        __typename: "Pokemon",
+      },
+      {
+        id: 30,
+        weight: 200,
+        height: 8,
+        __typename: "Pokemon",
+      },
+      {
+        id: 31,
+        weight: 600,
+        height: 13,
+        __typename: "Pokemon",
+      },
+      {
+        id: 32,
+        weight: 90,
+        height: 5,
+        __typename: "Pokemon",
+      },
+      {
+        id: 33,
+        weight: 195,
+        height: 9,
+        __typename: "Pokemon",
+      },
+      {
+        id: 34,
+        weight: 620,
+        height: 14,
+        __typename: "Pokemon",
+      },
+      {
+        id: 35,
+        weight: 75,
+        height: 6,
+        __typename: "Pokemon",
+      },
+      {
+        id: 36,
+        weight: 400,
+        height: 13,
+        __typename: "Pokemon",
+      },
+    ],
+    status: true,
+    message: "",
+    __typename: "PokemonList",
+  },
+};
