+++ src/test/mocks/fixtures/Pokemons/byOffset/offset54.ts
@@ -1,0 +1,134 @@
+export const offset54 = {
+  pokemons: {
+    count: 1302,
+    next: "http://localhost/api/v2/pokemon/?limit=9&offset=63",
+    previous: "http://localhost/api/v2/pokemon/?limit=9&offset=45",
+    nextOffset: 63,
+    prevOffset: 45,
+    params: { limit: 9, offset: 54 },
+    results: [
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/55.png",
+        id: 55,
+        name: "golduck",
+        url: "http://localhost/api/v2/pokemon/55/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/55.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/55.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/56.png",
+        id: 56,
+        name: "mankey",
+        url: "http://localhost/api/v2/pokemon/56/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/56.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/56.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/57.png",
+        id: 57,
+        name: "primeape",
+        url: "http://localhost/api/v2/pokemon/57/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/57.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/57.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/58.png",
+        id: 58,
+        name: "growlithe",
+        url: "http://localhost/api/v2/pokemon/58/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/58.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/58.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/59.png",
+        id: 59,
+        name: "arcanine",
+        url: "http://localhost/api/v2/pokemon/59/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/59.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/59.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/60.png",
+        id: 60,
+        name: "poliwag",
+        url: "http://localhost/api/v2/pokemon/60/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/60.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/60.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/61.png",
+        id: 61,
+        name: "poliwhirl",
+        url: "http://localhost/api/v2/pokemon/61/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/61.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/61.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/62.png",
+        id: 62,
+        name: "poliwrath",
+        url: "http://localhost/api/v2/pokemon/62/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/62.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/62.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/63.png",
+        id: 63,
+        name: "abra",
+        url: "http://localhost/api/v2/pokemon/63/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/63.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/63.svg",
+        __typename: "PokemonItem",
+      },
+    ],
+    extended_results: [
+      { id: 55, weight: 766, height: 17, __typename: "Pokemon" },
+      { id: 56, weight: 280, height: 5, __typename: "Pokemon" },
+      { id: 57, weight: 320, height: 10, __typename: "Pokemon" },
+      { id: 58, weight: 190, height: 7, __typename: "Pokemon" },
+      { id: 59, weight: 1550, height: 19, __typename: "Pokemon" },
+      { id: 60, weight: 124, height: 6, __typename: "Pokemon" },
+      { id: 61, weight: 200, height: 10, __typename: "Pokemon" },
+      { id: 62, weight: 540, height: 13, __typename: "Pokemon" },
+      { id: 63, weight: 195, height: 9, __typename: "Pokemon" },
+    ],
+    status: true,
+    message: "",
+    __typename: "PokemonList",
+  },
+};
