+++ src/test/mocks/fixtures/Pokemons/byOffset/offset9.ts
@@ -1,0 +1,182 @@
+export const offset9 = {
+  pokemons: {
+    count: 1302,
+    next: "http://localhost/api/v2/pokemon/?limit=9&offset=18",
+    previous: "http://localhost/api/v2/pokemon/?limit=9",
+    nextOffset: 18,
+    prevOffset: 0,
+    params: {
+      limit: 9,
+      offset: 9,
+    },
+    results: [
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/10.png",
+        id: 10,
+        name: "caterpie",
+        url: "http://localhost/api/v2/pokemon/10/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/10.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/10.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/11.png",
+        id: 11,
+        name: "metapod",
+        url: "http://localhost/api/v2/pokemon/11/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/11.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/11.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/12.png",
+        id: 12,
+        name: "butterfree",
+        url: "http://localhost/api/v2/pokemon/12/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/12.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/12.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/13.png",
+        id: 13,
+        name: "weedle",
+        url: "http://localhost/api/v2/pokemon/13/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/13.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/13.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/14.png",
+        id: 14,
+        name: "kakuna",
+        url: "http://localhost/api/v2/pokemon/14/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/14.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/14.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/15.png",
+        id: 15,
+        name: "beedrill",
+        url: "http://localhost/api/v2/pokemon/15/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/15.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/15.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/16.png",
+        id: 16,
+        name: "pidgey",
+        url: "http://localhost/api/v2/pokemon/16/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/16.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/16.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/17.png",
+        id: 17,
+        name: "pidgeotto",
+        url: "http://localhost/api/v2/pokemon/17/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/17.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/17.svg",
+        __typename: "PokemonItem",
+      },
+      {
+        image:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/18.png",
+        id: 18,
+        name: "pidgeot",
+        url: "http://localhost/api/v2/pokemon/18/",
+        artwork:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/18.png",
+        dreamworld:
+          "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/18.svg",
+        __typename: "PokemonItem",
+      },
+    ],
+    extended_results: [
+      {
+        id: 10,
+        weight: 29,
+        height: 3,
+        __typename: "Pokemon",
+      },
+      {
+        id: 11,
+        weight: 99,
+        height: 7,
+        __typename: "Pokemon",
+      },
+      {
+        id: 12,
+        weight: 320,
+        height: 11,
+        __typename: "Pokemon",
+      },
+      {
+        id: 13,
+        weight: 32,
+        height: 3,
+        __typename: "Pokemon",
+      },
+      {
+        id: 14,
+        weight: 100,
+        height: 6,
+        __typename: "Pokemon",
+      },
+      {
+        id: 15,
+        weight: 295,
+        height: 10,
+        __typename: "Pokemon",
+      },
+      {
+        id: 16,
+        weight: 18,
+        height: 3,
+        __typename: "Pokemon",
+      },
+      {
+        id: 17,
+        weight: 300,
+        height: 11,
+        __typename: "Pokemon",
+      },
+      {
+        id: 18,
+        weight: 395,
+        height: 15,
+        __typename: "Pokemon",
+      },
+    ],
+    status: true,
+    message: "",
+    __typename: "PokemonList",
+  },
+};
