+++ src/test/mocks/fixtures/Pokemons/index.ts
@@ -1,7 +1,1 @@
-export * from "./offset0";
-export * from "./offset9";
-export * from "./offset18";
-export * from "./offset27";
-export * from "./offset36";
-export * from "./offset45";
-export * from "./offset54";
+export * from "./byOffset";
