--- src/test/mocks/fixtures/Pokemons/offset18.ts
@@ -1,184 +1,0 @@
-export const offset18 = {
-  data: {
-    pokemons: {
-      count: 1302,
-      next: "http://localhost/api/v2/pokemon/?limit=9&offset=27",
-      previous: "http://localhost/api/v2/pokemon/?limit=9&offset=9",
-      nextOffset: 27,
-      prevOffset: 9,
-      params: {
-        limit: 9,
-        offset: 18,
-      },
-      results: [
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/19.png",
-          id: 19,
-          name: "rattata",
-          url: "http://localhost/api/v2/pokemon/19/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/19.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/19.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/20.png",
-          id: 20,
-          name: "raticate",
-          url: "http://localhost/api/v2/pokemon/20/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/20.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/20.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/21.png",
-          id: 21,
-          name: "spearow",
-          url: "http://localhost/api/v2/pokemon/21/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/21.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/21.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/22.png",
-          id: 22,
-          name: "fearow",
-          url: "http://localhost/api/v2/pokemon/22/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/22.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/22.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/23.png",
-          id: 23,
-          name: "ekans",
-          url: "http://localhost/api/v2/pokemon/23/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/23.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/23.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/24.png",
-          id: 24,
-          name: "arbok",
-          url: "http://localhost/api/v2/pokemon/24/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/24.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/24.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/25.png",
-          id: 25,
-          name: "pikachu",
-          url: "http://localhost/api/v2/pokemon/25/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/25.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/25.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/26.png",
-          id: 26,
-          name: "raichu",
-          url: "http://localhost/api/v2/pokemon/26/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/26.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/26.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/27.png",
-          id: 27,
-          name: "sandshrew",
-          url: "http://localhost/api/v2/pokemon/27/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/27.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/27.svg",
-          __typename: "PokemonItem",
-        },
-      ],
-      extended_results: [
-        {
-          id: 19,
-          weight: 35,
-          height: 3,
-          __typename: "Pokemon",
-        },
-        {
-          id: 20,
-          weight: 185,
-          height: 7,
-          __typename: "Pokemon",
-        },
-        {
-          id: 21,
-          weight: 20,
-          height: 3,
-          __typename: "Pokemon",
-        },
-        {
-          id: 22,
-          weight: 380,
-          height: 12,
-          __typename: "Pokemon",
-        },
-        {
-          id: 23,
-          weight: 69,
-          height: 20,
-          __typename: "Pokemon",
-        },
-        {
-          id: 24,
-          weight: 650,
-          height: 35,
-          __typename: "Pokemon",
-        },
-        {
-          id: 25,
-          weight: 60,
-          height: 4,
-          __typename: "Pokemon",
-        },
-        {
-          id: 26,
-          weight: 300,
-          height: 8,
-          __typename: "Pokemon",
-        },
-        {
-          id: 27,
-          weight: 120,
-          height: 6,
-          __typename: "Pokemon",
-        },
-      ],
-      status: true,
-      message: "",
-      __typename: "PokemonList",
-    },
-  },
-};
