--- src/test/mocks/fixtures/Pokemons/offset36.ts
@@ -1,184 +1,0 @@
-export const offset36 = {
-  data: {
-    pokemons: {
-      count: 1302,
-      next: "http://localhost/api/v2/pokemon/?limit=9&offset=45",
-      previous: "http://localhost/api/v2/pokemon/?limit=9&offset=27",
-      nextOffset: 45,
-      prevOffset: 27,
-      params: {
-        limit: 9,
-        offset: 36,
-      },
-      results: [
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/37.png",
-          id: 37,
-          name: "vulpix",
-          url: "http://localhost/api/v2/pokemon/37/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/37.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/37.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/38.png",
-          id: 38,
-          name: "ninetales",
-          url: "http://localhost/api/v2/pokemon/38/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/38.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/38.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/39.png",
-          id: 39,
-          name: "jigglypuff",
-          url: "http://localhost/api/v2/pokemon/39/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/39.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/39.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/40.png",
-          id: 40,
-          name: "wigglytuff",
-          url: "http://localhost/api/v2/pokemon/40/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/40.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/40.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/41.png",
-          id: 41,
-          name: "zubat",
-          url: "http://localhost/api/v2/pokemon/41/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/41.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/41.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/42.png",
-          id: 42,
-          name: "golbat",
-          url: "http://localhost/api/v2/pokemon/42/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/42.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/42.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/43.png",
-          id: 43,
-          name: "oddish",
-          url: "http://localhost/api/v2/pokemon/43/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/43.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/43.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/44.png",
-          id: 44,
-          name: "gloom",
-          url: "http://localhost/api/v2/pokemon/44/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/44.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/44.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/45.png",
-          id: 45,
-          name: "vileplume",
-          url: "http://localhost/api/v2/pokemon/45/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/45.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/45.svg",
-          __typename: "PokemonItem",
-        },
-      ],
-      extended_results: [
-        {
-          id: 37,
-          weight: 99,
-          height: 6,
-          __typename: "Pokemon",
-        },
-        {
-          id: 38,
-          weight: 199,
-          height: 11,
-          __typename: "Pokemon",
-        },
-        {
-          id: 39,
-          weight: 55,
-          height: 5,
-          __typename: "Pokemon",
-        },
-        {
-          id: 40,
-          weight: 120,
-          height: 10,
-          __typename: "Pokemon",
-        },
-        {
-          id: 41,
-          weight: 75,
-          height: 8,
-          __typename: "Pokemon",
-        },
-        {
-          id: 42,
-          weight: 550,
-          height: 16,
-          __typename: "Pokemon",
-        },
-        {
-          id: 43,
-          weight: 54,
-          height: 5,
-          __typename: "Pokemon",
-        },
-        {
-          id: 44,
-          weight: 86,
-          height: 8,
-          __typename: "Pokemon",
-        },
-        {
-          id: 45,
-          weight: 186,
-          height: 12,
-          __typename: "Pokemon",
-        },
-      ],
-      status: true,
-      message: "",
-      __typename: "PokemonList",
-    },
-  },
-};
