--- src/test/mocks/fixtures/Pokemons/offset45.ts
@@ -1,184 +1,0 @@
-export const offset45 = {
-  data: {
-    pokemons: {
-      count: 1302,
-      next: "http://localhost/api/v2/pokemon/?limit=9&offset=54",
-      previous: "http://localhost/api/v2/pokemon/?limit=9&offset=36",
-      nextOffset: 54,
-      prevOffset: 36,
-      params: {
-        limit: 9,
-        offset: 45,
-      },
-      results: [
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/46.png",
-          id: 46,
-          name: "paras",
-          url: "http://localhost/api/v2/pokemon/46/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/46.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/46.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/47.png",
-          id: 47,
-          name: "parasect",
-          url: "http://localhost/api/v2/pokemon/47/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/47.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/47.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/48.png",
-          id: 48,
-          name: "venonat",
-          url: "http://localhost/api/v2/pokemon/48/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/48.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/48.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/49.png",
-          id: 49,
-          name: "venomoth",
-          url: "http://localhost/api/v2/pokemon/49/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/49.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/49.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/50.png",
-          id: 50,
-          name: "diglett",
-          url: "http://localhost/api/v2/pokemon/50/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/50.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/50.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/51.png",
-          id: 51,
-          name: "dugtrio",
-          url: "http://localhost/api/v2/pokemon/51/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/51.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/51.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/52.png",
-          id: 52,
-          name: "meowth",
-          url: "http://localhost/api/v2/pokemon/52/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/52.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/52.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/53.png",
-          id: 53,
-          name: "persian",
-          url: "http://localhost/api/v2/pokemon/53/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/53.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/53.svg",
-          __typename: "PokemonItem",
-        },
-        {
-          image:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/54.png",
-          id: 54,
-          name: "psyduck",
-          url: "http://localhost/api/v2/pokemon/54/",
-          artwork:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/54.png",
-          dreamworld:
-            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/54.svg",
-          __typename: "PokemonItem",
-        },
-      ],
-      extended_results: [
-        {
-          id: 46,
-          weight: 54,
-          height: 3,
-          __typename: "Pokemon",
-        },
-        {
-          id: 47,
-          weight: 295,
-          height: 10,
-          __typename: "Pokemon",
-        },
-        {
-          id: 48,
-          weight: 300,
-          height: 10,
-          __typename: "Pokemon",
-        },
-        {
-          id: 49,
-          weight: 125,
-          height: 15,
-          __typename: "Pokemon",
-        },
-        {
-          id: 50,
-          weight: 8,
-          height: 2,
-          __typename: "Pokemon",
-        },
-        {
-          id: 51,
-          weight: 333,
-          height: 7,
-          __typename: "Pokemon",
-        },
-        {
-          id: 52,
-          weight: 42,
-          height: 4,
-          __typename: "Pokemon",
-        },
-        {
-          id: 53,
-          weight: 320,
-          height: 10,
-          __typename: "Pokemon",
-        },
-        {
-          id: 54,
-          weight: 196,
-          height: 8,
-          __typename: "Pokemon",
-        },
-      ],
-      status: true,
-      message: "",
-      __typename: "PokemonList",
-    },
-  },
-};
