+++ src/test/mocks/handlers/Pokemons.ts
@@ -1,26 +1,25 @@
+// Uses: https://mswjs.io/docs/network-behavior/graphql/#mocking-responses
+import { buildASTSchema, graphql as executeGraphQL } from "graphql";
 import { HttpResponse, graphql } from "msw";
+import Document from "schema.graphql";
 
 import * as paginatedResponse from "../fixtures/Pokemons";
+
+const schema = buildASTSchema(Document);
 
 const pokeapi = graphql.link("http://localhost:4000");
 
 export const Pokemons = [
-  pokeapi.query("Pokemons", ({ variables }) => {
-    switch (variables.offset) {
-      case 9:
-        return HttpResponse.json(paginatedResponse.offset9);
-      case 18:
-        return HttpResponse.json(paginatedResponse.offset18);
-      case 27:
-        return HttpResponse.json(paginatedResponse.offset27);
-      case 36:
-        return HttpResponse.json(paginatedResponse.offset36);
-      case 45:
-        return HttpResponse.json(paginatedResponse.offset45);
-      case 54:
-        return HttpResponse.json(paginatedResponse.offset54);
-      default:
-        return HttpResponse.json(paginatedResponse.offset0);
-    }
+  pokeapi.query("Pokemons", async ({ query, variables }) => {
+    const { errors, data } = await executeGraphQL({
+      schema,
+      source: query,
+      variableValues: variables,
+      rootValue: {
+        ...paginatedResponse[`offset${variables.offset}`],
+      },
+    });
+
+    return HttpResponse.json({ errors, data });
   }),
 ];
