+++ vite.config.ts
@@ -1,3 +1,4 @@
+import graphql from "@rollup/plugin-graphql";
 import react from "@vitejs/plugin-react-swc";
 import { defineConfig } from "vite";
 import tsconfigPaths from "vite-tsconfig-paths";
@@ -9,6 +10,7 @@ export default defineConfig(() => ({
       jsxImportSource: "@emotion/react",
     }),
     tsconfigPaths(),
+    graphql(),
   ],
   test: {
     environment: "jsdom",
